+++ config/configload/helper.go
@@ -6,8 +6,11 @@ import (
 
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/gohcl"
+	"github.com/hashicorp/hcl/v2/hclsyntax"
 
 	"github.com/avenga/couper/config"
+	"github.com/avenga/couper/config/sequence"
+	"github.com/avenga/couper/errors"
 	"github.com/avenga/couper/eval"
 )
 
@@ -69,7 +72,13 @@ func (h *helper) addBackend(block *hcl.Block) error {
 }
 
 func (h *helper) configureDefinedBackends() error {
-	for name, b := range h.defsBackends {
+	backendNames, err := h.resolveBackendDeps()
+	if err != nil {
+		return err
+	}
+
+	for _, name := range backendNames {
+		b := h.defsBackends[name]
 		be, err := PrepareBackend(h, "_init", "", &config.Backend{Name: name, Remain: b})
 		if err != nil {
 			return err
@@ -81,7 +90,7 @@ func (h *helper) configureDefinedBackends() error {
 
 		h.defsBackends[name] = be
 	}
-	return nil
+	return err
 }
 
 func (h *helper) configureACBackends() error {
@@ -105,4 +114,100 @@ func (h *helper) configureACBackends() error {
 		}
 	}
 	return nil
+}
+
+// resolveBackendDeps returns defined backends ordered by reference. Referenced ones need to be configured first.
+func (h *helper) resolveBackendDeps() (uniqueItems []string, err error) {
+	// collect referenced backends
+	refs := make(map[string][]string)
+	h.collectBackendDeps(refs)
+	// built up deps
+	refPtr := map[string]*sequence.Item{}
+	for name := range refs {
+		parent := &sequence.Item{Name: name}
+		refPtr[name] = parent
+	}
+
+	defer func() {
+		if p := recover(); p != nil { // since we use sequence related logic, replace wording due to backend context here
+			err = errors.Configuration.Message(strings.Replace(fmt.Sprintf("%s", p), "sequence ", "", 1))
+		}
+	}()
+
+	var defs sequence.List
+	for parent, ref := range refs {
+		for _, r := range ref {
+			p, _ := refPtr[parent]
+			if be, exist := refPtr[r]; exist {
+				p.Add(be)
+			} else {
+				p.Add(&sequence.Item{Name: r})
+			}
+			defs = append(defs, p)
+		}
+	}
+
+	items := sequence.Dependencies(defs)
+
+	// do not forget the other ones
+	var standalone []string
+	for def, _ := range h.defsBackends {
+		standalone = append(standalone, def)
+	}
+	items = append(items, standalone)
+
+	// unique by name /w score (sort?) // TODO: MAY refine with scoring of appearance
+	unique := make(map[string]int)
+	for _, seqItem := range items {
+		for _, name := range seqItem {
+			if _, exist := unique[name]; !exist {
+				unique[name] = 1
+				uniqueItems = append(uniqueItems, name)
+			} else {
+				unique[name]++
+			}
+		}
+	}
+
+	return uniqueItems, err
+}
+
+func (h *helper) collectBackendDeps(refs map[string][]string) {
+	for name, b := range h.defsBackends {
+		refs[name] = nil
+		content, _, _ := b.PartialContent(&hcl.BodySchema{
+			Blocks: []hcl.BlockHeaderSchema{{Type: oauth2}}},
+		)
+		oaBlocks := content.Blocks.OfType(oauth2)
+		for _, ob := range oaBlocks {
+			osb, ok := ob.Body.(*hclsyntax.Body)
+			if !ok {
+				continue
+			}
+
+			for _, be := range osb.Attributes {
+				if be.Name == backend {
+					val, _ := be.Expr.Value(envContext)
+					refs[name] = append(refs[name], val.AsString())
+					break
+				}
+			}
+
+			for _, block := range osb.Blocks {
+				if block.Type != backend {
+					continue
+				}
+				if len(block.Labels) > 0 {
+					refs[name] = append(refs[name], block.Labels[0])
+				}
+
+				for _, subBlock := range block.Body.Blocks {
+					if subBlock.Type == oauth2 {
+						h.collectBackendDeps(refs)
+						break
+					}
+				}
+			}
+		}
+	}
 }