+++ handler/transport/oauth2_req_auth.go
@@ -1,7 +1,6 @@
 package transport
 
 import (
-	"context"
 	"fmt"
 	"net/http"
 	"sync"
@@ -84,9 +83,8 @@ func (oa *OAuth2ReqAuth) RetryWithToken(req *http.Request, res *http.Response) (
 	oa.memStore.Del(oa.storageKey)
 
 	ctx := req.Context()
-	if retries, ok := ctx.Value(request.TokenRequestRetries).(uint8); !ok || retries < *oa.config.Retries {
-		ctx = context.WithValue(ctx, request.TokenRequestRetries, retries+1)
-
+	if retries, ok := ctx.Value(request.TokenRequestRetries).(*uint8); !ok || *retries < *oa.config.Retries {
+		*retries++ // increase ptr value instead of context value
 		req.Header.Del("Authorization")
 		err := oa.WithToken(req.WithContext(ctx)) // WithContext due to header manipulation
 		return true, err