+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Couper Changelog
 
-## [Unreleased](https://github.com/avenga/couper/compare/v1.9.1...master)
+## [Unreleased](https://github.com/avenga/couper/compare/v1.9.2...master)
 
 Unreleased changes are available as `avenga/couper:edge` container.
 
@@ -11,6 +11,16 @@ Unreleased changes are available as `avenga/couper:edge` container.
   * Endpoint path normalization to better match OpenAPI behavior ([#526](https://github.com/avenga/couper/pull/526))
 
 ---
+
+## [1.9.2](https://github.com/avenga/couper/releases/tag/v1.9.2)
+
+* **Fixed**
+  * configuration related panic while loading backends with [`oauth2` block](./docs/REFERENCE.md#oauth2-cc-block) which depends on other defined backends ([#524](https://github.com/avenga/couper/pull/524))
+  * erroneous retries for [`oauth2`](./docs/REFERENCE.md#oauth2-cc-block) backend authorization ([#529](https://github.com/avenga/couper/pull/529))
+    * with `retries = 0` ([#528](https://github.com/avenga/couper/pull/528))
+    * with `retries` > `0` and related origin configuration ([#529](https://github.com/avenga/couper/pull/529))
+  * race condition resulting in empty [`backends.<label>.health.state` variable](docs/REFERENCE.md#backends) ([#530](https://github.com/avenga/couper/pull/530))
+  * enabled json html escaping inherited from Go lib ([#531](https://github.com/avenga/couper/pull/531))
 
 ## [1.9.1](https://github.com/avenga/couper/releases/tag/v1.9.1)